generate_labels/davis_python/requirements.txt
cycler==0.10.0
decorator==4.0.11
easydict==1.7
enum34==1.1.6
functools32==3.2.3.post2
h5py==2.7.0
joblib==0.11
matplotlib==2.0.2
networkx==1.11
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
prettytable==0.7.2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.0
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
